+++ nodes/Zenlayer/Zenlayer.node.ts
@@ -28,17 +28,7 @@ export class Zenlayer implements INodeType {
 			},
 		},
 
-		inputs: `={{
-		(() => {
-			const resource = $parameter.resource;
-	  		const operation = $parameter.operation;
-			if (resource === 'text' && operation === 'message') {
-				return [{ type: 'main' }, { type: 'ai_tool', displayName: 'Tools' }];
-			}
-
-			return ['main'];
-		})()
-	}}`,
+		inputs: [NodeConnectionTypes.Main],
 		outputs: [NodeConnectionTypes.Main],
 
 		credentials: [