+++ CHANGELOG.md
@@ -3,6 +3,6 @@
 All notable changes to this project will be documented in this file. Dates are displayed in UTC.
 
 ## 0.1.6
-- Support MCP Client Tool
+- Remove redundant inputs
 - Support json_schema
 - Bug fixes and performance improvements